Project/requirements.txt
pandas
sqlalchemy
psycopg2-binary
jupyter
minio